git-svn-id: https://192.168.0.7/svn/team11/eRIMS_SONIC_New/trunk/eRIMS_Sonic@59574 8092c8c6-7ab6-f546-91b1-2980cf571e14

+++ eRims_SonicWeb/App_Code/SONIC_DAL/Building.cs
@@ -3283,7 +3283,7 @@ namespace ERIMS.DAL
             return db.ExecuteDataSet(dbCommand);
         }
 
-        public static DataSet BuildingByFacilityConstructionProjectId(int FK_Facility_Construction_PM)
+        public static DataSet BuildingByFacilityConstructionProjectId(decimal FK_Facility_Construction_PM)
         {
             Database db = DatabaseFactory.CreateDatabase();
             DbCommand dbCommand = db.GetStoredProcCommand("BuildingSelectByFacilityConstructionProjectId");
